RocksDB changes to support unity build: * Remove extra definition of TotalFileSize * Remove extra definition of ClipToRange * Move EncodedFileMetaData out from anonymous namespace * Move version_set Saver to its own namespace * Move some symbols into a named namespace * Move symbols out of anonymous namespace (prevents warning) * Make BloomHash inline
