+++ db/compaction.h
@@ -10,7 +10,11 @@
 #pragma once
 #include "db/version_set.h"
 
+#include <vector>
+
 namespace rocksdb {
+
+struct FileMetaData;
 
 class Version;
 class ColumnFamilyData;
@@ -154,5 +158,8 @@ class Compaction {
   // to pick up the next file to be compacted from files_by_size_
   void ResetNextCompactionIndex();
 };
+
+// Utility function
+extern uint64_t TotalFileSize(const std::vector<FileMetaData*>& files);
 
 }  // namespace rocksdb
