+++ util/bloom.cc
@@ -15,9 +15,6 @@
 namespace rocksdb {
 
 namespace {
-static uint32_t BloomHash(const Slice& key) {
-  return Hash(key.data(), key.size(), 0xbc9f1d34);
-}
 
 class BloomFilterPolicy : public FilterPolicy {
  private:
